requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.11
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==36.1.0
fastapi==0.115.8
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
playwright==1.50.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.1.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
redis==6.2.0
requests==2.32.3
rsa==4.9
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tenacity==9.0.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
pytest-asyncio